+++ BitFaster.Caching.UnitTests/Lru/ConcurrentLruTests.cs
@@ -1118,171 +1118,6 @@ namespace BitFaster.Caching.UnitTests.Lru
             removedItems[1].Reason.Should().Be(ItemRemovedReason.Trimmed);
         }
 
-        [Fact]
-        public async Task WhenSoakConcurrentGetCacheEndsInConsistentState()
-        {
-            for (int i = 0; i < 10; i++)
-            { 
-                await Threaded.Run(4, () => {
-                    for (int i = 0; i < 100000; i++)
-                    {
-                        lru.GetOrAdd(i + 1, i =>i.ToString());
-                    }
-                });
-
-                this.testOutputHelper.WriteLine($"{lru.HotCount} {lru.WarmCount} {lru.ColdCount}");
-                this.testOutputHelper.WriteLine(string.Join(" ", lru.Keys));
-
-                // allow +/- 1 variance for capacity
-                lru.Count.Should().BeCloseTo(9, 1);
-                RunIntegrityCheck();
-            }
-        }
-
-        [Fact]
-        public async Task WhenSoakConcurrentGetAsyncCacheEndsInConsistentState()
-        {
-            for (int i = 0; i < 10; i++)
-            {
-                await Threaded.RunAsync(4, async () => {
-                    for (int i = 0; i < 100000; i++)
-                    {
-                        await lru.GetOrAddAsync(i + 1, i => Task.FromResult(i.ToString()));
-                    }
-                });
-
-                this.testOutputHelper.WriteLine($"{lru.HotCount} {lru.WarmCount} {lru.ColdCount}");
-                this.testOutputHelper.WriteLine(string.Join(" ", lru.Keys));
-
-                // allow +/- 1 variance for capacity
-                lru.Count.Should().BeCloseTo(9, 1);
-                RunIntegrityCheck();
-            }
-        }
-
-        [Fact]
-        public async Task WhenSoakConcurrentGetWithArgCacheEndsInConsistentState()
-        {
-            for (int i = 0; i < 10; i++)
-            {
-                await Threaded.Run(4, () => {
-                    for (int i = 0; i < 100000; i++)
-                    {
-                        // use the arg overload
-                        lru.GetOrAdd(i + 1, (i, s) => i.ToString(), "Foo");
-                    }
-                });
-
-                this.testOutputHelper.WriteLine($"{lru.HotCount} {lru.WarmCount} {lru.ColdCount}");
-                this.testOutputHelper.WriteLine(string.Join(" ", lru.Keys));
-
-                // allow +/- 1 variance for capacity
-                lru.Count.Should().BeCloseTo(9, 1);
-                RunIntegrityCheck();
-            }
-        }
-
-        [Fact]
-        public async Task WhenSoakConcurrentGetAsyncWithArgCacheEndsInConsistentState()
-        {
-            for (int i = 0; i < 10; i++)
-            {
-                await Threaded.RunAsync(4, async () => {
-                    for (int i = 0; i < 100000; i++)
-                    {
-                        // use the arg overload
-                        await lru.GetOrAddAsync(i + 1, (i, s) => Task.FromResult(i.ToString()), "Foo");
-                    }
-                });
-
-                this.testOutputHelper.WriteLine($"{lru.HotCount} {lru.WarmCount} {lru.ColdCount}");
-                this.testOutputHelper.WriteLine(string.Join(" ", lru.Keys));
-
-                // allow +/- 1 variance for capacity
-                lru.Count.Should().BeCloseTo(9, 1);
-                RunIntegrityCheck();
-            }
-        }
-
-        [Fact]
-        public async Task WhenSoakConcurrentGetAndRemoveCacheEndsInConsistentState()
-        {
-            for (int i = 0; i < 10; i++)
-            {
-                await Threaded.Run(4, () => {
-                    for (int i = 0; i < 100000; i++)
-                    {
-                        lru.TryRemove(i + 1);
-                        lru.GetOrAdd(i + 1, i => i.ToString());
-                    }
-                });
-
-                this.testOutputHelper.WriteLine($"{lru.HotCount} {lru.WarmCount} {lru.ColdCount}");
-                this.testOutputHelper.WriteLine(string.Join(" ", lru.Keys));
-
-                RunIntegrityCheck();
-            }
-        }
-
-        [Fact]
-        public async Task WhenSoakConcurrentGetAndRemoveKvpCacheEndsInConsistentState()
-        {
-            for (int i = 0; i < 10; i++)
-            {
-                await Threaded.Run(4, () => {
-                    for (int i = 0; i < 100000; i++)
-                    {
-                        lru.TryRemove(new KeyValuePair<int, string>(i + 1, (i + 1).ToString()));
-                        lru.GetOrAdd(i + 1, i => i.ToString());
-                    }
-                });
-
-                this.testOutputHelper.WriteLine($"{lru.HotCount} {lru.WarmCount} {lru.ColdCount}");
-                this.testOutputHelper.WriteLine(string.Join(" ", lru.Keys));
-
-                RunIntegrityCheck();
-            }
-        }
-
-        [Fact]
-        public async Task WhenSoakConcurrentGetAndUpdateCacheEndsInConsistentState()
-        {
-            for (int i = 0; i < 10; i++)
-            {
-                await Threaded.Run(4, () => {
-                    for (int i = 0; i < 100000; i++)
-                    {
-                        lru.TryUpdate(i + 1, i.ToString());
-                        lru.GetOrAdd(i + 1, i => i.ToString());
-                    }
-                });
-
-                this.testOutputHelper.WriteLine($"{lru.HotCount} {lru.WarmCount} {lru.ColdCount}");
-                this.testOutputHelper.WriteLine(string.Join(" ", lru.Keys));
-
-                RunIntegrityCheck();
-            }
-        }
-
-        [Fact]
-        public async Task WhenSoakConcurrentGetAndAddCacheEndsInConsistentState()
-        {
-            for (int i = 0; i < 10; i++)
-            {
-                await Threaded.Run(4, () => {
-                    for (int i = 0; i < 100000; i++)
-                    {
-                        lru.AddOrUpdate(i + 1, i.ToString());
-                        lru.GetOrAdd(i + 1, i => i.ToString());
-                    }
-                });
-
-                this.testOutputHelper.WriteLine($"{lru.HotCount} {lru.WarmCount} {lru.ColdCount}");
-                this.testOutputHelper.WriteLine(string.Join(" ", lru.Keys));
-
-                RunIntegrityCheck();
-            }
-        }
 
         private void Warmup()
         {
@@ -1297,10 +1132,7 @@ namespace BitFaster.Caching.UnitTests.Lru
             lru.GetOrAdd(-9, valueFactory.Create);
         }
 
-        private void RunIntegrityCheck()
-        {
-            new ConcurrentLruIntegrityChecker<int, string, LruItem<int, string>, LruPolicy<int, string>, TelemetryPolicy<int, string>>(this.lru).Validate();
-        }
+
     }
 
     public class ConcurrentLruIntegrityChecker<K, V, I, P, T>
